compiler warnings in imgui_plot.cpp

+++ include/Mahi/Gui/imgui_plot.hpp
@@ -143,7 +143,7 @@ inline void PlotItemRollPoint(PlotItem& item, double x, double y, double span = 
 }
 /// Pushes a point into an item data set as if it were a circular buffer of #max_points size
 inline void PlotItemBufferPoint(PlotItem& item, double x, double y, int max_points) {
-    if (item.data.size() < max_points) 
+    if (item.data.size() < static_cast<std::size_t>(max_points)) 
         item.data.push_back(ImVec2(static_cast<float>(x),static_cast<float>(y)));
     else {
         item.data[item.data_begin] = ImVec2(static_cast<float>(x),static_cast<float>(y));
